Move disabling functions to cpu_stark.rs
